move Home component
